Moved unnamed-variable-namer to ast2ram utils.
